+++ src/App.vue
@@ -1,5 +1,21 @@
 <template>
   <app-sidebar></app-sidebar>
+  <alert-ui
+    v-model="showVersionAlert"
+    class="fixed w-full max-w-2xl new-version-alert rounded-b-none md:rounded-b-lg md:mb-4 bottom-0"
+    style="z-index: 9999"
+    dismissible
+  >
+    Snapcode v{{ latestVersion }} is released!! To see what's new open
+    <a
+      href="https://github.com/Kholid060/snapcode/blob/master/CHANGELOG.md"
+      rel="noopener"
+      target="_blank"
+      class="border-b-2 border-white"
+    >
+      changelog.
+    </a>
+  </alert-ui>
   <div class="lg:pl-64">
     <router-view :route="routeWithModal"></router-view>
   </div>
@@ -23,7 +39,9 @@ export default {
     const theme = useTheme();
     const router = useRouter();
 
+    const latestVersion = ref('');
     const isRetrieved = ref(false);
+    const showVersionAlert = ref(false);
 
     theme.setTheme(localStorage.getItem('theme') || 'dark');
 
@@ -37,6 +55,17 @@ export default {
       try {
         await store.dispatch('retrieveData');
         await retrieveBackupData();
+
+        const releaseCache = localStorage.getItem('snapcode-version') || '';
+        const appLatestVersion = import.meta.env.VITE_SNAPCODE_VERSION;
+        const isFirstTime = JSON.parse(localStorage.getItem('firstTime') || 'true');
+
+        if (releaseCache !== appLatestVersion && !isFirstTime) {
+          showVersionAlert.value = true;
+          latestVersion.value = appLatestVersion;
+
+          localStorage.setItem('snapcode-version', appLatestVersion);
+        }
 
         store.commit('updateState', {
           key: 'isRetrieved',
@@ -53,8 +82,16 @@ export default {
 
     return {
       isRetrieved,
+      latestVersion,
       routeWithModal,
+      showVersionAlert,
     };
   },
 };
 </script>
+<style>
+.new-version-alert {
+  transform: translateX(-50%);
+  left: 50%;
+}
+</style>