+++ src/components/pages/home/gists/GistsList.vue
@@ -51,8 +51,8 @@
 </template>
 <script>
 import { reactive, onMounted, computed } from 'vue';
-import { useStore } from 'vuex';
-import { Folder, File } from '~/models';
+import { useStorage } from '~/composable';
+import { Folder } from '~/models';
 
 export default {
   props: {
@@ -67,7 +67,7 @@ export default {
   },
   emits: ['close'],
   setup(props, { emit }) {
-    const store = useStore();
+    const storage = useStorage();
 
     const state = reactive({
       query: '',
@@ -135,15 +135,10 @@ export default {
 
       snippets.forEach(({ status, value }) => {
         if (status === 'fulfilled') {
-          File.$update({
+          storage.model('files').update({
             data: value,
           });
         }
-      });
-
-      store.commit('updateState', {
-        key: 'isDataChanged',
-        value: true,
       });
 
       state.loadingImport = false;