+++ src/composable/index.js
@@ -1,3 +1,6 @@
 export { default as useGroupTooltip } from './groupTooltip';
 export { default as useDialog } from './dialog';
 export { default as useTheme } from './theme';
+export { default as useStorage } from './storage';
+export { default as useComponentId } from './componentId';
+export { default as useIntersect } from './intersect';